+++ service/controller/key/types.go
@@ -3,6 +3,7 @@ package key
 // AppSpec is used to define app custom resources.
 type AppSpec struct {
 	App     string
+	AppName string
 	Catalog string
 	Chart   string
 	// Whether app is installed for clusterapi clusters only.
@@ -10,6 +11,9 @@ type AppSpec struct {
 	// ConfigMapName overrides the name, otherwise the cluster values configmap
 	// is used.
 	ConfigMapName string
+	// InCluster determines if the app CR should use in cluster. Otherwise the
+	// cluster kubeconfig is specified.
+	InCluster bool
 	// Whether app is installed for legacy clusters only.
 	LegacyOnly      bool
 	Namespace       string